- extra queues to deal with interruptible sleep are dead end
looks like extra queus, which deal with interruptible sleep
are not worth the effort. the only information the turnstile_block()
requires is whether it should setup interruptible sleep. If
interruptible slepp is not set up, then scheduler will not
try to wake up sleeping process. That is my understanding, which
I should clarify with ?kettenis, mpi?