- we need to make in_ouraddr() more picky (or paranoid) on destination   address it accepts as ours.
the idea is to introduce a paranoid mode, which enforces match
of IP destination address found in IP header with IP address
bound to interface.

The releax mode is used in conjuction with 'fowarding' being
enabled. In this case in_ouraddr() falls back to current behavior,
it is happy with packet if its destination address matches any IP
address bound to host.