- friction between anchors/ruleset and tables.
let us consider pf.conf with rules as follows:

	table <dup> { 192.168.1.0/24, 192.168.2.0/24 }

	anchor foo all
	load anchor "foo" from "pf-foo.conf"

content of pf-foo.conf reads like that:

	table <foo> { 172.16.0.0/16, 172.17.0.0/16 }
	pass from <dup> to <foo>

loading pf.conf above results in configuration as follows:

	lumpy# pfctl -a "*" -sr
	anchor "foo" all {
	  pass from <dup> to <foo> flags S/SA
	}

what has happened? pf.conf file provides a definition of table
<dup> which is later referred by pass rule in anchor/ruleset foo.
pf-foo.conf defines tbale <foo> which is also referred by the
pass rule in anchor/ruleset foo. So let's display tables:

	lumpy# pfctl -sT
	dup
	lumpy# pfctl -a foo -sT
	foo

there are two tables: dup which is a part of main ruleset/anchor
and table foo which belongs to ruleset/anchor foo.

Let's step back and try to simplify configuration by putting
everything into a single file:

	table <dup> { 192.168.1.0/24, 192.168.2.0/24 }
	anchor "foo" {
		table <foo> { 172.16.0.0/16, 172.17.0.0/16 }
		pass from <dup> to <foo>
	}

rules above should be the same as what we've seen earlier
main ruleset provides definition of table dup and anchor
rule "foo". We also inline definition of anchor/ruleset
foo. The ruleset foo is same as pf-foo.conf. This is what
happens when we load simplified version:

	lumpy# pfctl -f pf.conf
	/tmp/pf.conf:3: syntax error
	/tmp/pf.conf:5: syntax error
	pfctl: Syntax error in config file: pf rules not loaded

The 'Syntax error' is easily fixed by this commit. The change moves
us to new error:

	lumpy# /home/sashan/src.sashan/sbin/pfctl/pfctl -f pf.conf
	pf.conf:3: cannot define table foo: Device busy
	pf.conf:5: syntax error
	pfctl: Syntax error in config file: pf rules not loaded

this time we face error which comes from pf(4) itself right here
line 1906:

	1888 pfr_ina_define(struct pfr_table *tbl, struct pfr_addr *addr, int size,
	1889     int *nadd, int *naddr, u_int32_t ticket, int flags)
	1890 {
	1891         struct pfr_ktableworkq   tableq;
	1892         struct pfr_kentryworkq   addrq;
	1893         struct pfr_ktable       *kt, *rt, *shadow, key;
	1894         struct pfr_kentry       *p;
	1895         struct pfr_addr          ad;
	1896         struct pf_ruleset       *rs;
	1897         int                      i, rv, xadd = 0, xaddr = 0;
	1898
	1899         ACCEPT_FLAGS(flags, PFR_FLAG_DUMMY | PFR_FLAG_ADDRSTOO);
	1900         if (size && !(flags & PFR_FLAG_ADDRSTOO))
	1901                 return (EINVAL);
	1902         if (pfr_validate_table(tbl, PFR_TFLAG_USRMASK,
	1903             flags & PFR_FLAG_USERIOCTL))
	1904                 return (EINVAL);
	1905         rs = pf_find_ruleset(tbl->pfrt_anchor);
	1906         if (rs == NULL || !rs->topen || ticket != rs->tticket)
	1907                 return (EBUSY);
	1908         tbl->pfrt_flags |= PFR_TFLAG_INACTIVE;
	1909         SLIST_INIT(&tableq);

The problem is pfctl calls pfr_ina_define() before transaction
is opened. This fix is not that straightforward.

However this is not the single friction/complication we deal with
when it comes to ruleset and tables. Let's see what happens when
we fix pf.conf file to:

	table <dup> { 192.168.1.0/24, 192.168.2.0/24 }
	anchor "foo" {
		table <foo> { 172.16.0.0/16, 172.17.0.0/16 }
		pass from <dup> to <foo>
	}

File above loads with no issues:

	lumpy# pfctl -f /tmp/pf.conf
	lumpy# pfctl -a "*" -sr
	anchor "foo" all {
	  pass from <dup> to <foo> flags S/SA
	}

Let's see what tables are defined:

	lumpy# pfctl -sT
	dup
	lumpy# pfctl -a foo -sT

Now we want to add a missing definition for table foo:

	lumpy# pfctl -t foo -T add 172.16.0.0/16
	pfctl: Table does not exist

Let's do another test. This time we will use config as follows:

	table <dup> { 192.168.1.0/24, 192.168.2.0/24 }
	pass from <dup> to any

We load it and kill table:

	lumpy# pfctl -f /tmp/pf.conf
	lumpy# pfctl -sr
	pass from <dup> to any flags S/SA
	lumpy# pfctl -t dup -T kill
	1 table deleted.

now firewall configuration is more or less busted:

	lumpy# pfctl -t dup -T add 172.16.0.0/16
	pfctl: Table does not exist