X509_NAME_print: remove lie about multiple lines being used
OpenSSL commit 92ada7cc (2007) removed some dead code with flawed logic
attempting to print multiple lines if the line exceeded 80 characters.
Said flawed logic was there since the start of the git history importing
SSLeay 0.8.1b in 1998 and never worked. Rumor has it that it did work prior
to that. Be that as it may, it's just wrongly documented since Henson added
the docs in commit 0711be16 (2002).

Prompted by OpenSSL issue #18004 by davidben
https://github.com/quictls/quictls/pull/168
https://github.com/quictls/quictls/issues/75