On i386, llvm >= 17 warns with -Wsync-alignment when using a sync builtin with a 64-bit value that isn't 64-bit aligned.
The Intel SDM volume 3, strongly suggests to do this for performance
reasons.  My reading is this isn't required to be atomic.

Make atomic64_t 64-bit aligned and change arguments to avoid the warning.
And make a local change to the inteldrm code to set alignment in a case
where we use a sync builtin when linux doesn't (cmpxchg64()).

discussed with tb@ robert@ kettenis@, ok tb@