Make wsmouse(4) and wstpad filterops mp-safe.
Please note, filt_wseventread() is the only thread that runs
simultaneously with the rest of the wscons(4). This thread only compare
`ws_get' with `ws_put'. wscons(4) drivers only increment `ws_put' index
so at driver side we only need to protect this increment to ensure
`ws_put' is consistent with WSEVENT_QSIZE. I wrapped `ws_put' and
`ws_get' access with `ws_mtx' mutex(9) but this is not required.

ok kirill