Inline BN_reciprocal() in its only caller
This is simpler, doesn't need an auxiliary function of dubious value,
avouds an auxiliary variable and gets rid of a bunch of comments that
are hard to make sense of.

This doesn't bother to invalidate recp->shift since on error you should
not be reusing the RECP_CTX without reinitializing it.

ok jsing