add bpfxattach for addding extra bpf interfaces to net interfaces
bpf interfaces are identified by name and data link type. bpfattach
takes an ifnet pointer and uses the interface name from if_xname
when setting up a bpf interface for a network interface. this makes
it hard to provide a way to look at a subset of packets on an
interface with the same dlt as the whole interface.

bpfxattach lets me attach a bpf interface with a different name to
network interfaces. im using this to attach bpf interfaces rings
on some multiq interfaces, which lets me fan out packet processing
via bpf in userland.

ok claudio@