Fix scope of the sa_mpls sockaddr variable.
The sa_mpls is used outside of the block via info.rti_info[RTAX_SRC] so
it needs to have the same scope as info.

Yesterday the planets and the moon were in perfect alignment to trigger
another bug in route(8) because of this and I could observe all of its
glory first hand for the first time after 7 years.
OK bluhm@ tb@