takey_print: don't emit empty line between comments and URIs
It was probably an oversight in RFC 8630 that TALs don't allow an empty
line between the comment section and the URI section. Or maybe a comment
of only one or two lines was intended rather than the ugly blob we now
have in arin.tal. Be that as it may, the output is supposed to be a TAL
and it should conform to the RFC.

ok claudio