let the device know that we are moving to INIT state in qwx_stop()
This disables the VIF and other firmware-internal state before sending
the STOP command to turn wifi off.

Second part of fixes for pool corruption caused by the device writing
to mbufs on Rx rings while the driver is trying to stop the interface.

Tested by myself on z13 and by kettenis on x13s