allocate a uid/gid for lldpd to run with as an _lldpd user
the net/lldpd port also uses _lldpd, and will now share with base.
depending on which appears first (base or port _lldpd) determines
which uid/gid will be used. neither base or port lldpds store
anything on disk, the user and group are only used to drop privileges
to and run as.

discussed with and ok sthen@ deraadt@